chore(refactor): Done project refactoring for own_engine searcher
