+++ src/forms/documents/document.rs
@@ -4,7 +4,7 @@ use chrono::{DateTime, Datelike, NaiveDateTime, Timelike, Utc};
 use datetime::{deserialize_dt, serialize_dt};
 use derive_builder::Builder;
 use serde::{Deserialize, Deserializer, Serialize};
-use utoipa::{IntoParams, ToSchema};
+use utoipa::ToSchema;
 
 #[derive(Deserialize, Serialize, Builder, Default, Clone, ToSchema)]
 pub struct Document {
@@ -215,14 +215,26 @@ impl TestExample<Document> for Document {
 #[derive(Builder, Clone, Deserialize, Serialize, ToSchema)]
 pub struct OcrMetadata {
     #[schema(example = "c643c506-f5c3-4262-991d-bbe847035499")]
-    pub job_id: String,
-    pub text: String,
+    job_id: String,
+    text: String,
     #[schema(example = 1)]
-    pub pages_count: i32,
+    pages_count: i32,
     #[schema(example = "Коносамент")]
-    pub doc_type: String,
+    doc_type: String,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub artifacts: Option<Vec<Artifacts>>,
+    artifacts: Option<Vec<Artifacts>>,
+}
+
+impl OcrMetadata {
+    pub fn get_job_id(&self) -> &str {
+        self.job_id.as_str()
+    }
+    pub fn get_pages_count(&self) -> i32 {
+        self.pages_count
+    }
+    pub fn get_artifacts(&self) -> Option<&Vec<Artifacts>> {
+        self.artifacts.as_ref()
+    }
 }
 
 #[derive(Builder, Clone, Deserialize, Serialize, ToSchema)]
@@ -233,6 +245,12 @@ pub struct Artifacts {
     pub group_json_name: String,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub group_values: Option<Vec<GroupValue>>,
+}
+
+impl Artifacts {
+    pub fn builder() -> ArtifactsBuilder {
+        ArtifactsBuilder::default()
+    }
 }
 
 #[derive(Builder, Clone, Deserialize, Serialize, ToSchema)]
@@ -259,154 +277,4 @@ where
 #[derive(Builder, Clone, Default, Deserialize, Serialize, ToSchema)]
 pub struct HighlightEntity {
     pub content: Vec<String>,
-}
-
-#[derive(Builder, Clone, Default, Deserialize, Serialize, ToSchema)]
-pub struct DocumentPreview {
-    #[schema(example = "98ac9896be35f47fb8442580cd9839b4")]
-    id: String,
-    #[schema(example = "test_document.txt")]
-    name: String,
-    #[serde(
-        serialize_with = "serialize_dt",
-        deserialize_with = "deserialize_dt",
-        skip_serializing_if = "Option::is_none"
-    )]
-    #[schema(example = "2024-04-03T13:51:32Z")]
-    created_at: Option<DateTime<Utc>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    quality_recognition: Option<i32>,
-    #[schema(example = 35345)]
-    file_size: i32,
-    #[schema(example = "Test Folder")]
-    location: String,
-    #[schema(example = "test_folder")]
-    folder_id: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    preview_properties: Option<Vec<Artifacts>>,
-}
-
-impl DocumentPreview {
-    pub fn get_id(&self) -> &str {
-        self.id.as_str()
-    }
-
-    pub fn get_name(&self) -> &str {
-        self.name.as_str()
-    }
-
-    pub fn get_folder_id(&self) -> &str {
-        self.folder_id.as_str()
-    }
-
-    pub fn get_location(&self) -> &str {
-        self.location.as_str()
-    }
-}
-
-impl TestExample<DocumentPreview> for DocumentPreview {
-    fn test_example(_val: Option<&str>) -> DocumentPreview {
-        let created = NaiveDateTime::default()
-            .with_year(2024)
-            .unwrap()
-            .with_month(4)
-            .unwrap()
-            .with_day(3)
-            .unwrap()
-            .with_hour(13)
-            .unwrap()
-            .with_minute(51)
-            .unwrap()
-            .with_second(32)
-            .unwrap()
-            .and_utc();
-
-        DocumentPreviewBuilder::default()
-            .id("98ac9896be35f47fb8442580cd9839b4".to_string())
-            .name("test_document.txt".to_string())
-            .created_at(Some(created))
-            .quality_recognition(Some(10000))
-            .file_size(35345)
-            .location("test_folder".to_string())
-            .preview_properties(
-                vec![ArtifactsBuilder::default()
-                    .group_name("Information of TN".to_string())
-                    .group_json_name("tn_info".to_string())
-                    .group_values(
-                        vec![GroupValueBuilder::default()
-                            .name("Date of TN".to_string())
-                            .json_name("date_of_tn".to_string())
-                            .group_type("string".to_string())
-                            .value(Some("2023-10-29".to_string()))
-                            .build()
-                            .unwrap()]
-                        .into(),
-                    )
-                    .build()
-                    .unwrap()]
-                .into(),
-            )
-            .build()
-            .unwrap()
-    }
-}
-
-impl From<Document> for DocumentPreview {
-    fn from(value: Document) -> Self {
-        DocumentPreviewBuilder::default()
-            .id(value.document_md5)
-            .name(value.document_name)
-            .created_at(value.document_created)
-            .quality_recognition(Some(10000))
-            .file_size(value.document_size)
-            .location(value.folder_id)
-            .preview_properties(Some(Vec::default()))
-            .build()
-            .unwrap()
-    }
-}
-
-#[derive(Builder, Clone, Default, Deserialize, Serialize, IntoParams, ToSchema)]
-pub struct MoveDocumetsForm {
-    document_ids: Vec<String>,
-    location: String,
-    src_folder_id: String,
-}
-
-impl MoveDocumetsForm {
-    pub fn get_folder_id(&self) -> &str {
-        self.location.as_str()
-    }
-
-    pub fn get_src_folder_id(&self) -> &str {
-        self.src_folder_id.as_str()
-    }
-
-    pub fn get_document_ids(&self) -> &[String] {
-        self.document_ids.as_slice()
-    }
-}
-
-impl TestExample<MoveDocumetsForm> for MoveDocumetsForm {
-    fn test_example(_value: Option<&str>) -> MoveDocumetsForm {
-        MoveDocumetsFormBuilder::default()
-            .location("Test Folder".to_string())
-            .src_folder_id("unrecognized".to_string())
-            .document_ids(vec!["98ac9896be35f47fb8442580cd9839b4".to_string()])
-            .build()
-            .unwrap()
-    }
-}
-
-#[derive(Clone, Default, Deserialize, Serialize, IntoParams, ToSchema)]
-pub struct AnalyseDocumentsForm {
-    pub document_ids: Vec<String>,
-}
-
-impl TestExample<AnalyseDocumentsForm> for AnalyseDocumentsForm {
-    fn test_example(_value: Option<&str>) -> AnalyseDocumentsForm {
-        AnalyseDocumentsForm {
-            document_ids: vec!["98ac9896be35f47fb8442580cd9839b4".to_string()],
-        }
-    }
 }
