+++ src/forms/schemas/vector.rs
@@ -1,0 +1,58 @@
+use crate::forms::schemas::FieldType;
+use crate::forms::schemas::SchemaFieldType;
+
+use serde_derive::Serialize;
+
+#[derive(Serialize)]
+pub struct DocumentVectorSchema {
+    mappings: DocumentVectorMappings,
+}
+
+#[derive(Serialize)]
+struct DocumentVectorMappings {
+    properties: DocumentVectorProperties,
+}
+
+#[derive(Serialize)]
+struct DocumentVectorProperties {
+    folder_id: SchemaFieldType,
+    document_id: SchemaFieldType,
+    text_vector: TextVectorSchema,
+}
+
+#[derive(Serialize)]
+struct TextVectorSchema {
+    #[serde(rename(serialize = "type"))]
+    field_type: FieldType,
+    properties: TextVectorProperties,
+}
+
+#[derive(Serialize)]
+struct TextVectorProperties {
+    vector: SchemaFieldType,
+    text_chunk: SchemaFieldType,
+}
+
+impl Default for DocumentVectorSchema {
+    fn default() -> Self {
+        let text_vec_properties = TextVectorProperties {
+            vector: SchemaFieldType::new(FieldType::DenseVector),
+            text_chunk: SchemaFieldType::new(FieldType::Text),
+        };
+
+        let doc_vec_properties = DocumentVectorProperties {
+            folder_id: SchemaFieldType::new(FieldType::String),
+            document_id: SchemaFieldType::new(FieldType::String),
+            text_vector: TextVectorSchema {
+                field_type: FieldType::Nested,
+                properties: text_vec_properties,
+            },
+        };
+
+        DocumentVectorSchema {
+            mappings: DocumentVectorMappings {
+                properties: doc_vec_properties,
+            },
+        }
+    }
+}
