+++ src/services/elastic/searcher/mod.rs
@@ -1,1 +1,2 @@
+pub(crate) mod helper;
 pub mod searcher;
