+++ src/services/mod.rs
@@ -4,5 +4,5 @@ pub mod init;
 mod notifier;
 pub mod own_engine;
 pub mod redis_cache;
-pub mod searcher;
+pub mod service;
 pub mod swagger;
