+++ src/services/own_engine/searcher/mod.rs
@@ -1,0 +1,2 @@
+mod helper;
+pub(super) mod searcher;
